Labs/Lab1/requirements.txt
prettytable >= 1.0.1
python-dateutil
